restructure wordBank and utils
