+++ src/wordBank/parts/adjective.js
