+++ src/wordBank/parts/interjection.js
