+++ src/wordBank/parts/nounSubject.js
