+++ src/wordBank/utils.js
